apps/api/requirements.txt
fastapi==0.128.2
uvicorn[standard]==0.40.0
python-docx==1.2.0
sqlalchemy==2.0.44
pydantic-settings==2.12.0
httpx==0.28.1
PyJWT==2.12.1
pytest==9.0.2
langchain==1.2.13
langchain-openai==1.1.12
langchain-core==1.2.22
langchain-text-splitters==1.1.1
pinecone-client==6.0.0
psycopg[binary]==3.2.12
